Moving the webhook extension to be instance-based so that we are not dealing with global state.
