+++ src/SmtpPilot/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "SmtpPilot": {
       "commandName": "Project",
-      "commandLineArgs": "-i any -w http://localhost:53085/smtppilot/",
+      "commandLineArgs": "-i any -w http://localhost:5000/smtppilot/",
       "environmentVariables": {
         "COMPlus_DebugWriteToStdErr": "1"
       }
